+++ facebook_business/adobjects/techprovider/examples/deleteaccount.py
@@ -1,0 +1,33 @@
+# Copyright 2014 Facebook, Inc.
+
+# You are hereby granted a non-exclusive, worldwide, royalty-free license to
+# use, copy, modify, and distribute this software in source code or binary
+# form for use in connection with the web services and APIs provided by
+# Facebook.
+
+# As with any software that integrates with the Facebook platform, your use
+# of this software is subject to the Facebook Developer Principles and
+# Policies [http://developers.facebook.com/policy/]. This copyright notice
+# shall be included in all copies or substantial portions of the software.
+
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+# DEALINGS IN THE SOFTWARE.
+
+from facebook_business.adobjects.techprovider.thirdpartyaccountmanager import (
+    ThirdPartyAccountManager,
+)
+
+tech_provider_business_id = "<tech_provider_business_id>"
+app_id = "<app_id>"
+system_account_token = "<system_account_token>"
+
+third_party_account_manager = ThirdPartyAccountManager(
+    app_id, tech_provider_business_id, system_account_token
+)
+accound_id_to_delete = input("Enter the account id to delete: ")
+third_party_account_manager.delete_account(accound_id_to_delete)